First pass at hiding JS dependencies from implementation
Create script tags dynamically inside the constructor, pulling the ACE JavaScript libraries
from the relative package path.

Don't create an instance of the ACE editor until the dynamically created script has
loaded.

Add a guard in the content setter to prevent calling methods on a JS proxy that might not
be up yet.

Huge thanks to @santiaago for pairing on this.