First pass at localStorage for ICE projects
Something of a pain because it needs to be ordered (LIFO), need to be
able to lookup by key (title), needs to map into JSON for localStorage. So we're
serializing to a JSON list of HashMap, where the title is in the HashMap, but
also serves as the lookup key.

So far, we have insert and update working.

Paired with @korishev