Attempting to generate keyboard event using JS
http://jsbin.com/awenaq/3 was the closest I came, but even then Dart didn't recognise the keyCode attribute that was added.

Tried https://github.com/termi/DOM-Keyboard-Event-Level-3-polyfill as part of the HTML and calling it via js.context['...'].callMethod that didn't work

They all either didn't have the keyCode field or the keyCode was set to zero.

Even Object.defineProperty failed to define and/or set the keyCode fields with the right values.