First pass at storage capable of ignoring snapshots
Current thinking is to store the periodic snapshots along with regular projects, but
add a flag that can be ignored.

Our first attempt is to switch the in-memory storage to a LinkedHashMap. This seems
promising.  Need to fix a few tests still and double-check assumptions elsewhere.

Open questions: does the projects getter return a list or LinkedHashMap.

Paired with @terrasea

[#87]