Tested the non-removal of lock when editor is locked
Removed a bunch of debug code.

We saw the lock dissappear from time-to-time while reloading, but were unable
to reliably reproduce. The editorLock.existing should never change values so
this current approach seems solid. Still, it's a little weird that we saw
this on occasion.

Hopefully just a cache issue. Since we do not see this when compiled to JS,
we further hope that we are not checking in a race condition. Fingers crossed!

Paired with @terrasea

#90