Identify test package hash as source of trouble
The test package sends a bunch of info to individual tests in
the window location hash, so our mucking with it interfers
with the tests. Need to find a way around that, but still
need to be able to test things like the removal of hash strings
when copy dialog closes.

Paired with @kgrz

#102