Dupe preview frame to use when online
Trying to suppress appcache messages when preview frame is loaded. The
idea is to make a regular network connection to get preview frame when online
(bypassing appcache). And using the cached preview frame when offline.

Gonna be a pain to keep two identical copies in sync, but hopefully the
preview frame rarely changes.